backend/requirements.txt
fastapi
uvicorn
python-multipart
soundfile
requests
python-dotenv
edge-tts
torch
transformers
phonemizer
protobuf